reorder code in mutiple files
