+++ app/src/Flash/SPFS.cpp
@@ -1,7 +1,6 @@
 #include "SPFS.h"
 #include "flash.h"
 #include <cstring>
-#include <cstring>
 
 std::shared_ptr<SPFS::Directory> SPFS::getRootDirectory(int start_offset, int end_offset) {
   if(_fs_header == nullptr) {
@@ -69,64 +68,6 @@ bool SPFS::formatDisk(const void *address, size_t size) {
   return Flash::erase(address, size) == 0;
 }
 
-std::shared_ptr<SPFS::Directory> SPFS::Directory::createDirectory(const std::string& name){
-  auto new_dir = _fs->createDirectory(shared_from_this(), name);
-  if(new_dir == nullptr){
-    return nullptr;
-  }
-
-  std::vector<uint8_t> buffer(FS_BLOCK_SIZE);
-  if(Flash::read(buffer, getHeader()) < (int)buffer.size()) {
-    return nullptr;
-  }
-
-  DirectoryHeader *dirheader = reinterpret_cast<DirectoryHeader *>(buffer.data());
-
-  auto contentHeaders = reinterpret_cast<DirectoryContentHeader *>(reinterpret_cast<uint8_t*>(dirheader) + (dirheader->name_size_content_offset >> 8));
-  auto max_count = (int)((FS_BLOCK_SIZE - (dirheader->name_size_content_offset >> 8)) / sizeof(DirectoryContentHeader));
-
-  int current = 0;
-  while(contentHeaders[current].type != 0xFFFF && current < max_count) {
-    current++;
-  }
-
-  if(current >= max_count) {
-    //TODO: Implement directory extension
-    return nullptr; // No space for new content
-  }
-
-  contentHeaders[current].type = MAGIC_SUBDIRMARKER; // Directory type
-  contentHeaders[current].block_offset = (int16_t)(((uintptr_t)new_dir->getHeader() - (uintptr_t)getHeader()) / FS_BLOCK_SIZE);
-
-  if(Flash::write(buffer, getHeader()) < (int)buffer.size()) {
-    return nullptr;
-  }
-
-  return new_dir;
-}
-
-std::vector<std::shared_ptr<SPFS::Directory>> SPFS::Directory::getSubdirectories() {
-  std::vector<std::shared_ptr<SPFS::Directory>> subdirs;
-  // Implementation to populate subdirs from the directory content
-
-  auto contentHeaders = reinterpret_cast<const DirectoryContentHeader *>(reinterpret_cast<const uint8_t*>(getHeader()) + (getHeader()->name_size_content_offset >> 8));
-  auto max_count = (int)((FS_BLOCK_SIZE - (getHeader()->name_size_content_offset >> 8)) / sizeof(DirectoryContentHeader));
-
-  for(int i = 0; i < max_count; i++) {
-    if(contentHeaders[i].type == MAGIC_SUBDIRMARKER) { // Directory type
-      auto dir_address = reinterpret_cast<const uint8_t*>(getHeader()) + contentHeaders[i].block_offset * FS_BLOCK_SIZE;
-      auto subdir = _fs->openDirectory(dir_address, shared_from_this());
-      if(subdir != nullptr) {
-        subdirs.push_back(subdir);
-      }
-    }else if(contentHeaders[i].type == MAGIC_ENDMARKER) {
-      break; // End of content
-    }
-  }
-
-  return subdirs;
-}
-
 std::shared_ptr<SPFS::DirectoryInternal> SPFS::openDirectory(const void* address, std::shared_ptr<SPFS::Directory> parent) {
   auto *dirheader = reinterpret_cast<const DirectoryHeader *>(address);
   if(dirheader->magic != MAGIC_DIR_NUMBER) {
@@ -166,11 +107,6 @@ std::shared_ptr<SPFS::DirectoryInternal> SPFS::createDirectory(const void* addre
   }
 
   return std::make_shared<SPFS::DirectoryInternal>(this, parent, reinterpret_cast<const DirectoryHeader *>(address));
-}
-
-const std::string SPFS::Directory::getName() const {
-  const char* name_ptr = reinterpret_cast<const char*>(_header) + sizeof(SPFS::DirectoryHeader);
-  return std::string(name_ptr, _header->name_size_content_offset & 0x00FF);
 }
 
 std::shared_ptr<SPFS::Directory> SPFS::createNewFileSystem(const void *address, size_t size, const std::string& fs_name, const std::string& root_dir_name) {
